fix(den): streamline org invite signup (#1291)
